Continue refactor
Remove some old data structures in favor of preset
